% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the safeRview Shiny App}
\usage{
launch_app()
}
\description{
Opens the interactive Shiny dashboard for exploring adverse event data.
}
\examples{
\dontrun{
  launch_app()
}
}
