% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ae.R
\name{plot_ae_barchart}
\alias{plot_ae_barchart}
\title{Plot Top Adverse Events as a Bar Chart}
\usage{
plot_ae_barchart(ae_summary_table, top_n = 10)
}
\arguments{
\item{ae_summary_table}{The data.frame created by \code{summarize_ae_data()}.}

\item{top_n}{The number of adverse events to display.}
}
\value{
A ggplot object.
}
\description{
Creates a ggplot2 bar chart of the top N most frequent adverse events,
faceted by treatment arm.
}
